Switch to airline, and use gutentags
Gutentags provides unobtrusive, project-specific ctags functionality, making
tags useful.

Because tags are useful, it's nice to be able to display them in the status
line. Tagbar provides this functionality, in conjunction with phpctags. The next
trick is integration with the statusline software in use.

Powerline, however, makes doing so so opaque as to be impossible to
decipher. As such, switched to airline, which is lightweight, fast, and offers
easy to understand documentation covering configuration.

CtrlP is something I've been meaning to investigate, and since it integrates
well with gutentags, seemed a good fit. This required removing phpdoc.vim,
however, as that overrides Ctrl-P as well; since I found the functionality
useless since adopting snipMate, I'm fine with that.

Finally, discovered the `;` override in php.vim, which will make `f` useful
again, finally!